requirements.txt
torch>=2.2
numpy>=1.24
pillow>=10.0
openpyxl>=3.1
fastapi>=0.110
uvicorn>=0.29
python-multipart>=0.0.9
open-clip-torch==2.26.1
timm>=0.9.7
requests>=2.31
transformers>=4.45.0
bitsandbytes>=0.43.0
peft>=0.12.0
accelerate>=0.34.0
safetensors>=0.4.3
